feature(add-docker): Add docker and refactor mock-api
This adds a Dockerfile for the reactjs-todo application and the
e2e/mock-api

the mock-api got refactored to leverage the webpack config for
reloading and deployment

Added a docker-compose file which describes how to start each service
for later use in okteto
