+++ e2e/mock-api/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "mock-api",
+  "version": "1.0.0",
+  "description": "",
+  "main": "./index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
+    "express": "4.17.2",
+    "superagent": "7.0.1",
+    "uuid": "8.3.2"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
