+++ e2e/mock-api/src/app/app.certs.js
@@ -1,7 +1,7 @@
 /*
  * @forgerock/javascript-sdk
  *
- * app.certs.mjs
+ * app.certs.js
  *
  * Copyright (c) 2020 ForgeRock. All rights reserved.
  * This software may be modified and distributed under the terms
