+++ e2e/mock-api/src/app/constants.js
@@ -1,7 +1,7 @@
 /*
  * @forgerock/javascript-sdk
  *
- * constants.mjs
+ * constants.js
  *
  * Copyright (c) 2020 ForgeRock. All rights reserved.
  * This software may be modified and distributed under the terms
