+++ e2e/mock-api/src/app/env.config.js
@@ -1,7 +1,7 @@
 /*
  * @forgerock/javascript-sdk
  *
- * env.config.ts
+ * env.config.js
  *
  * Copyright (c) 2020 ForgeRock. All rights reserved.
  * This software may be modified and distributed under the terms
@@ -59,7 +59,7 @@ if (env.LIVE) {
 
 export const APP_PORT = ports.app;
 export const AM_PORT = amPort;
-export const MOCK_PORT = ports.mock;
+export const MOCK_PORT = process.env.PORT || ports.mock;
 
 export const AM_URL = `${amUrl}:${amPort}${paths.am}`;
 export const BASE_URL = `${origins.app}:${ports.app}`;
