+++ e2e/mock-api/src/app/response.registration.js
@@ -1,7 +1,7 @@
 /*
  * @forgerock/javascript-sdk
  *
- * response.registration.mjs
+ * response.registration.js
  *
  * Copyright (c) 2020 ForgeRock. All rights reserved.
  * This software may be modified and distributed under the terms
