+++ e2e/mock-api/src/app/routes.resource.js
@@ -1,7 +1,7 @@
 /*
  * @forgerock/javascript-sdk
  *
- * routes.resource.mjs
+ * routes.resource.js
  *
  * Copyright (c) 2020 ForgeRock. All rights reserved.
  * This software may be modified and distributed under the terms
@@ -10,12 +10,12 @@
 
 import { env } from 'process';
 import request from 'superagent';
-import { session } from './app.auth.mjs';
-import { key, cert } from './app.certs.mjs';
-import { AM_URL, AM_PORT, FORGEOPS, REALM_PATH } from './env.config.copy.mjs';
-import { authByTreeResponse, authByTxnResponse, createStepUpUrl } from './responses.mjs';
-import { baz } from './routes.auth.mjs';
-import wait from './wait.mjs';
+import { session } from './app.auth';
+import { key, cert } from './app.certs';
+import { AM_URL, AM_PORT, FORGEOPS, REALM_PATH } from './env.config';
+import { authByTreeResponse, authByTxnResponse, createStepUpUrl } from './responses';
+import { baz } from './routes.auth';
+import wait from './wait';
 
 async function authorization(req, res, next) {
   if (env.NODE_ENV === 'LIVE' && req.hostname !== FORGEOPS) {
