+++ samples/reactjs-todo/project.json
@@ -8,7 +8,6 @@
       "outputs": ["{options.outputPath}"],
       "options": {
         "outputPath": "dist/samples/reactjs-todo",
-        "generatePackageJson": true,
         "index": "samples/reactjs-todo/public/index.html",
         "main": "samples/reactjs-todo/client/index.js",
         "tsConfig": "samples/reactjs-todo/tsconfig.app.json",
@@ -17,20 +16,30 @@
           "includePaths": ["samples/reactjs-todo/public", "samples/reactjs-todo/client"]
         }
       },
-      "configurations": {}
+      "configurations": {
+        "production": {
+          "optimization": true,
+          "watch": false,
+          "webpackConfig": "samples/reactjs-todo/webpack.config.js",
+          "outputPath": "dist/samples/reactjs-todo",
+          "index": "samples/reactjs-todo/public/index.html",
+          "main": "samples/reactjs-todo/client/index.js",
+          "tsConfig": "samples/reactjs-todo/tsconfig.app.json"
+        }
+      }
     },
     "serve": {
       "executor": "@nrwl/web:dev-server",
       "defaultConfiguration": "development",
       "options": {
-        "buildTarget": "reactjs-todo:build",
-        "open": true,
-        "watch": true,
-        "ssl": true
+        "buildTarget": "reactjs-todo:build"
       },
       "configurations": {
-        "development": {},
-        "production": {}
+        "development": {
+          "open": true,
+          "watch": true,
+          "ssl": true
+        }
       }
     },
     "test": {
