+++ samples/todo-api/project.json
@@ -35,6 +35,15 @@
         "watch": true
       }
     },
+    "deploy": {
+      "executor": "@nrwl/workspace:run-commands",
+      "options": {
+        "commands": [
+          "nx build todo-api --prod",
+          "docker build -f ./samples/todo-api/Dockerfile . -t todo-api"
+        ]
+      }
+    },
     "lint": {
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
