+++ samples/todo-api/src/index.js
@@ -7,7 +7,6 @@
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
  */
-
 import cors from 'cors';
 import express from 'express';
 import cookieParser from 'cookie-parser';
